refactor: command service 도입
