+++ src/main/java/com/onmoim/server/common/image/repository/ImageRepository.java
@@ -2,8 +2,6 @@ package com.onmoim.server.common.image.repository;
 
 import com.onmoim.server.common.image.entity.Image;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
 
-@Repository
 public interface ImageRepository extends JpaRepository<Image, Long> {
-} 
+}
